refactor decoders
