+++ internal/pkg/application/decoder/elsys/decoders.elsys.go
@@ -1,4 +1,4 @@
-package decoder
+package elsys
 
 import (
 	"context"
