+++ internal/pkg/application/decoder/enviot/decoders.enviot.go
@@ -1,4 +1,4 @@
-package decoder
+package enviot
 
 import (
 	"context"
