+++ internal/pkg/application/decoder/milesight/decoders.milesight.go
@@ -1,4 +1,4 @@
-package decoder
+package milesight
 
 import (
 	"context"
