+++ internal/pkg/application/decoder/qalcosonic/decoder.qalcosonic.go
@@ -1,4 +1,4 @@
-package decoder
+package qalcosonic
 
 import (
 	"bytes"
