+++ internal/pkg/application/decoder/sensative/decoder.presence.go
@@ -1,4 +1,4 @@
-package decoder
+package sensative
 
 import (
 	"context"
