+++ internal/pkg/application/decoder/sensative/decoders.sensative.go
@@ -1,4 +1,4 @@
-package decoder
+package sensative
 
 import (
 	"context"
