+++ internal/pkg/application/decoder/sensefarm/decoders.sensefarm.go
@@ -1,4 +1,4 @@
-package decoder
+package sensefarm
 
 import (
 	"bytes"
